# libs/api-client/specs/event-stack.openapi.yaml
openapi: 3.0.3
info:
  title: NX Playground event-stack API
  description: >
    Spine contract dumped from Nest Events + Orders (not TypeSpec /console/events).
    Frontends talk to live Nest (:3001) or stateful api-mock (:3011) with the same paths.
  version: '1.0.0'
servers:
  - url: http://localhost:3001/api
    description: live Nest + Prisma
  - url: http://localhost:3011/api
    description: stateful contract mock
tags:
  - name: events
  - name: orders
  - name: tickets
  - name: payments
paths:
  /events:
    get:
      tags: [events]
      summary: Get all events
      operationId: listEvents
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: number
        - in: query
          name: limit
          schema:
            type: number
      responses:
        '200':
          description: Paginated events
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventListResponse'
    post:
      tags: [events]
      summary: Create new event
      operationId: createEvent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEventDto'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '400':
          description: Invalid input
  /events/{id}:
    get:
      tags: [events]
      summary: Get event by ID
      operationId: getEvent
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Event
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '404':
          description: Event not found
    put:
      tags: [events]
      summary: Update event
      operationId: updateEvent
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEventDto'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '404':
          description: Event not found
    delete:
      tags: [events]
      summary: Delete event
      operationId: deleteEvent
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted
        '404':
          description: Event not found
  /orders:
    get:
      tags: [orders]
      summary: List orders
      operationId: listOrders
      parameters:
        - in: query
          name: userId
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: number
        - in: query
          name: limit
          schema:
            type: number
      responses:
        '200':
          description: Paginated orders
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderListResponse'
    post:
      tags: [orders]
      summary: Create an order for an event
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderDto'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Invalid event or user
  /orders/{id}:
    get:
      tags: [orders]
      summary: Get order by ID
      operationId: getOrder
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Order not found
  /orders/{id}/confirm:
    post:
      tags: [orders]
      summary: Confirm an order
      operationId: confirmOrder
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Confirmed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Order not found
  /orders/{id}/tickets:
    get:
      tags: [tickets]
      summary: List tickets issued for an order
      operationId: listOrderTickets
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Tickets
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketListResponse'
        '404':
          description: Order not found
  /tickets/{id}:
    get:
      tags: [tickets]
      summary: Get ticket by ID
      operationId: getTicket
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Ticket
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ticket'
        '404':
          description: Ticket not found
  /tickets/{id}/verify:
    get:
      tags: [tickets]
      summary: Verify a ticket for check-in
      operationId: verifyTicket
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Verification
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketVerifyResponse'
        '404':
          description: Ticket not found
  /tickets/{id}/check-in:
    post:
      tags: [tickets]
      summary: Check in a ticket
      operationId: checkInTicket
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Checked in
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ticket'
        '400':
          description: Ticket cannot be checked in
        '404':
          description: Ticket not found
  /payments/intents:
    get:
      tags: [payments]
      summary: List payment intents for an order
      operationId: listPaymentIntents
      parameters:
        - in: query
          name: orderId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Intents
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentIntentListResponse'
        '404':
          description: Order not found
    post:
      tags: [payments]
      summary: Create a payment intent (no card data in this API)
      operationId: createPaymentIntent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePaymentIntentDto'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentIntent'
        '404':
          description: Order not found
  /payments/intents/{id}:
    get:
      tags: [payments]
      summary: Get a payment intent
      operationId: getPaymentIntent
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Intent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentIntent'
        '404':
          description: Payment intent not found
  /payments/webhook:
    post:
      tags: [payments]
      summary: Payment webhook / callback (idempotent; no card data)
      operationId: postPaymentWebhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentWebhookDto'
      responses:
        '200':
          description: Applied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentWebhookResponse'
        '404':
          description: Payment intent not found
  /payments/mock-complete/{id}:
    get:
      tags: [payments]
      summary: Mock checkout chooser when ECPay sandbox keys are absent
      operationId: mockCompletePayment
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: query
          name: outcome
          schema:
            type: string
            enum: [paid, failed]
      responses:
        '200':
          description: HTML chooser
        '302':
          description: Redirect to portal order page after mock outcome
        '400':
          description: Not a mock intent
        '404':
          description: Payment intent not found
components:
  schemas:
    CreateEventDto:
      type: object
      required: [title, startDate, endDate]
      additionalProperties: false
      properties:
        title:
          type: string
        description:
          type: string
        location:
          type: string
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        maxAttendees:
          type: number
        status:
          type: string
          enum: [draft, published]
        formId:
          type: string
    Event:
      type: object
      required: [id, title, startDate, endDate, status, createdAt, updatedAt]
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        location:
          type: string
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        maxAttendees:
          type: number
        status:
          type: string
          enum: [draft, published, cancelled]
        formId:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    EventListResponse:
      type: object
      required: [items, total, page, limit, totalPages]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Event'
        total:
          type: number
        page:
          type: number
        limit:
          type: number
        totalPages:
          type: number
    CreateOrderDto:
      type: object
      required: [eventId]
      additionalProperties: false
      properties:
        eventId:
          type: string
        userId:
          type: string
        status:
          type: string
          enum: [pending, confirmed, cancelled]
        data:
          type: object
          additionalProperties: true
    Order:
      type: object
      required: [id, eventId, userId, status, data, createdAt, updatedAt]
      properties:
        id:
          type: string
        eventId:
          type: string
        userId:
          type: string
        status:
          type: string
          enum: [pending, confirmed, cancelled]
        data:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    OrderListResponse:
      type: object
      required: [items, total, page, limit, totalPages]
      properties:
        totalPages:
          type: number
    Ticket:
      type: object
      required: [id, orderId, eventId, type, status, createdAt, updatedAt]
      properties:
        id:
          type: string
        orderId:
          type: string
        eventId:
          type: string
        type:
          type: string
        status:
          type: string
          enum: [issued, used, cancelled]
        checkedInAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    TicketListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Ticket'
    TicketVerifyResponse:
      type: object
      required: [ticket, event, order, isValid, verificationTime]
      properties:
        ticket:
          $ref: '#/components/schemas/Ticket'
        event:
          $ref: '#/components/schemas/Event'
        order:
          $ref: '#/components/schemas/Order'
        isValid:
          type: boolean
        verificationTime:
          type: string
          format: date-time
    CreatePaymentIntentDto:
      type: object
      required: [orderId]
      additionalProperties: false
      properties:
        orderId:
          type: string
    PaymentIntent:
      type: object
      required:
        - id
        - orderId
        - provider
        - status
        - merchantTradeNo
        - amount
        - checkoutUrl
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
        orderId:
          type: string
        provider:
          type: string
          enum: [mock, ecpay-sandbox]
        status:
          type: string
          enum: [created, paid, failed]
        merchantTradeNo:
          type: string
        amount:
          type: number
        checkoutUrl:
          type: string
          description: Redirect target. Never a card form hosted by this API.
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PaymentIntentListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PaymentIntent'
    PaymentWebhookDto:
      type: object
      properties:
        merchantTradeNo:
          type: string
        MerchantTradeNo:
          type: string
        rtnCode:
          description: '1 = paid, anything else = failed'
          oneOf:
            - type: string
            - type: number
        RtnCode:
          description: ECPay form field
          oneOf:
            - type: string
            - type: number
        CheckMacValue:
          type: string
    PaymentWebhookResponse:
      type: object
      required: [intent, replayed]
      properties:
        intent:
          $ref: '#/components/schemas/PaymentIntent'
        replayed:
          type: boolean